Add posts.html.tpl and prepare.php for preparing the posts page which will be used as an alternative page for those who run their browsers with JavaScript disabled.
Albeit the links on that page will point to raw markdown files
instead of the traditional Web pages, which may suck for some people
but you can always use a third-party tool, or use conversion tools
to generate HTML from Markdown and browse locally.

Links in the markdown files may point to `~/` which is a derivative
from the Unix/Linux eco-system that points to the HOME directory of
the logged in user.  I use this same approach to reference the HOME
directory of a particular post, those resources should be located at
`/cdn/$id/` so you can do a bit of `sed` magic and replace those
occurences if you want.

The reason I did it that way is to preserve bandwidth, as writing
the entire link is not very bandwidth friendly, but also the page
domain may change in the future and that has to be reflected
dynamically, which is at the moment done through JavaScript.

Anyways, bye.